requirements.txt
absl-py==2.3.0
addict==2.4.0
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
blis==0.7.11
Brotli==1.0.9
brotlicffi==1.0.9.2
cachetools==5.5.2
catalogue==1.0.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
decorator==4.4.2
docx==0.2.4
docx2pdf==0.1.8
docx2txt==0.9
easydict==1.9
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==2.2.5
flask-cors==6.0.1
flatbuffers==2.0.7
fonttools==4.55.3
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
huggingface-hub==0.33.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==6.11.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
json5==0.9.25
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
librosa==0.9.1
llvmlite==0.43.0
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.9
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
ml_dtypes==0.5.1
mmcv-full==1.6.1
mmdet==2.25.1
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.13
namex==0.1.0
narwhals==1.31.0
networkx==3.2.1
nltk==3.5
numba==0.60.0
numpy==1.24.3
open_clip_torch==2.29.0
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.16.0
packaging==23.2
pandas==2.0.3
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==10.2.0
pip==25.1.1
plac==1.1.3
platformdirs==4.3.8
plotly==6.0.1
pooch==1.8.2
preshed==3.0.10
proglog==0.1.12
protobuf==3.20.3
pyarrow==20.0.0
pycocotools==2.0.10
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PyQt6==6.7.1
PyQt6_sip==13.9.1
pyresparser==1.0.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rpds-py==0.25.1
sacremoses==0.1.1
safetensors==0.5.3
scenedetect==0.6.4
scikit-learn==1.6.1
scipy==1.13.1
setuptools==80.9.0
sip==6.10.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
soundfile==0.13.1
spacy==2.3.8
srsly==1.0.7
streamlit==1.47.0
sympy==1.13.3
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
terminaltables==3.1.10
thinc==7.4.2
threadpoolctl==3.6.0
timm==1.0.16
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.2.2
torchvision==0.17.2
tornado==6.2
tqdm==4.67.1
transformers==4.17.0
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
unicodedata2==15.1.0
urllib3==1.26.9
validators==0.35.0
wasabi==0.10.1
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.17.2
yapf==0.43.0
zipp==3.23.0